Switch to Tera for templating
The built in filters in Handlebars are very limited. I don't want to
decide on, implement, and maintain a dozen new filters. I was originally
planning to use Rhai scripting, but the user experience seems a lot
worse for flexibility gains no one may ever use. Tera seems to provide a
nicely chosen set of filters that I think provides a good balance of
power and convenience.