Add CPU block
I originally didn't intend to implement a CPU block since I figured
it could be produced with something like the `awk` command for memory
usage in the examples.

It turns out to that there really isn't a simple command for getting CPU
usage. The standard approach is to poll /proc/stat. In principle you
could accomplish this with a `command` block and a suitable binary.
A shell script that passes `top` or `mpstat` output through `awk` would
work, but I'm trying to avoid requiring shell scripts, and expecting
people to write a custom binary is a bit over the top, so this seemed
worth implementing!