Improve Inotify block startup behavior
Before, if there was an error reading the file on first read, the block
would fail to start for what was quite possibly a recoverable error.

By switching to doing all file reads asynchronously all errors get
treated the same way. Failures will be printed to stderr, but we'll
continue to watch the file.