Fix memory examples to show usage.
The examples were calculating free memory, which isn't actually very
useful. In practice available memory is the thing we actually care
about since it includes things like disk cache that's sitting in ram
because we're not using it, but which could be cleared immediately.

Also, I chose colors as if the block was showing ram usage, which
I think is the more intuitive thing to show, so I've updated the formula
to show usage as `100 - percent available`.